Replace MongoDB with ArangoDB
