+++ Dockerfile.development
@@ -27,15 +27,9 @@ COPY testnet/supervisord.conf /etc/supervisor/conf.d/supervisord.conf
 RUN mkdir -p /ot-node/init
 RUN ln -s /ot-node/init /ot-node/current
 
-RUN apt-key adv --keyserver hkp://keyserver.ubuntu.com:80 --recv 7F0CEB10
-RUN echo "deb http://repo.mongodb.org/apt/ubuntu "$(lsb_release -sc)"/mongodb-org/3.0 multiverse" | tee /etc/apt/sources.list.d/mongodb-org-3.0.list
-RUN apt-get update
-RUN apt-get install -y --allow-unauthenticated mongodb-org
-RUN apt install -y mongodb-server
-
 # Add files
 COPY . /ot-node/current
-RUN service arangodb3 start && service mongodb start && cd /ot-node/current && npm install && npm run setup -- --configDir=/ot-node/data
+RUN service arangodb3 start && cd /ot-node/current && npm install && npm run setup -- --configDir=/ot-node/data
 
 RUN wget https://github.com/papertrail/remote_syslog2/releases/download/v0.20/remote_syslog_linux_amd64.tar.gz
 RUN tar xzf ./remote_syslog_linux_amd64.tar.gz && cd remote_syslog && cp ./remote_syslog /usr/local/bin
