+++ package.json
@@ -138,7 +138,6 @@
     "minimist": "^1.2.0",
     "mkdirp": "^0.5.1",
     "moment": "^2.22.2",
-    "mongodb": "^3.6.3",
     "ms": "^2.1.1",
     "nedb": "^1.8.0",
     "neo4j-driver": "^1.6.2",
