+++ scripts/setup.js
@@ -16,7 +16,6 @@ const fs = require('fs');
 const { execSync } = require('child_process');
 const argv = require('minimist')(process.argv.slice(2));
 const rc = require('rc');
-const MongoDB = require('../modules/Database/MongoDB');
 
 // Check for arguments sanity.
 if (argv.all && argv.configDir) {
@@ -148,12 +147,5 @@ arangoDbs.forEach((database) => {
         process.abort();
     });
 });
-
-
-const db = new MongoDB(
-    'origintrail',
-    'staging_data',
-    null,
-);
 
 db.reinitialize();
