+++ testnet/supervisord.conf
@@ -24,12 +24,6 @@ autorestart=true
 stdout_logfile=/dev/null
 stdout_logfile_maxbytes=0
 
-[program:mongodb]
-command=systemctl start mongod
-autorestart=true
-stdout_logfile=/dev/null
-stdout_logfile_maxbytes=0
-
 [program:remote_syslog]
 command=remote_syslog -D
 stdout_logfile=/dev/null
